refactor: remove access-token-service and move it to authentication-service
